requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.3.0
babel==2.16.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3

pip~=24.3.1
geopy~=2.4.1
geographiclib~=2.0